Refactor extract the election building step
